requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain>=0.1.10,<0.2.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
rich>=13.0.0